+++ log4k/src/jsTest/kotlin/saschpe/log4k/LoggedTest.js.kt
@@ -1,4 +1,0 @@
-package saschpe.log4k
-
-internal actual val expectedListTag: String = "ArrayList"
-internal actual val expectedMapTag: String = "HashMap"