+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-05-02
+- Provide FileLogger for persistent logging to a file path with rotation and retention policy support.
+
 ## [1.2.5] - 2024-03-23
 - Provide javadoc artifacts for Sonatype Maven Central
 
@@ -13,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Allow empty log messages if you only want to create a log entry about a function being called.
 - Add more Apple ARM64 platforms: macOS, tvOS, watchOS
+- Provide FileLogger in addition to ConsoleLogger
 - Dependency update:
   - [Kotlin 1.9.23](https://kotlinlang.org/docs/whatsnew19.html)
   - [Gradle-8.7](https://docs.gradle.org/8.7/release-notes.html)